剥离更多 API 到 FlapApi 类
